split schema
