+++ src/services/database/schema/interface.ts
@@ -1,0 +1,22 @@
+import { sql } from 'drizzle-orm';
+import { int, sqliteTable, text } from 'drizzle-orm/sqlite-core';
+
+// maybe support multiple interfaces in the future
+export const wgInterface = sqliteTable('interface_table', {
+  id: int().primaryKey({ autoIncrement: true }),
+  device: text().notNull(),
+  port: int().notNull(),
+  privateKey: text('private_key').notNull(),
+  publicKey: text('public_key').notNull(),
+  ipv4Cidr: text('ipv4_cidr').notNull(),
+  ipv6Cidr: text('ipv6_cidr').notNull(),
+  mtu: int().notNull(),
+  enabled: int({ mode: 'boolean' }).notNull(),
+  createdAt: text('created_at')
+    .notNull()
+    .default(sql`(CURRENT_TIMESTAMP)`),
+  updatedAt: text('updated_at')
+    .notNull()
+    .default(sql`(CURRENT_TIMESTAMP)`)
+    .$onUpdate(() => sql`(CURRENT_TIMESTAMP)`),
+});
