+++ src/services/database/schema/oneTimeLinks.ts
@@ -1,0 +1,15 @@
+import { sql } from 'drizzle-orm';
+import { int, sqliteTable, text } from 'drizzle-orm/sqlite-core';
+
+export const oneTimeLinks = sqliteTable('one_time_links_table', {
+  id: int().primaryKey({ autoIncrement: true }),
+  oneTimeLink: text('one_time_link').notNull(),
+  expiresAt: text('expires_at').notNull(),
+  createdAt: text('created_at')
+    .notNull()
+    .default(sql`(CURRENT_TIMESTAMP)`),
+  updatedAt: text('updated_at')
+    .notNull()
+    .default(sql`(CURRENT_TIMESTAMP)`)
+    .$onUpdate(() => sql`(CURRENT_TIMESTAMP)`),
+});
