+++ src/services/database/schema/users.ts
@@ -1,0 +1,19 @@
+import { sql } from 'drizzle-orm';
+import { int, sqliteTable, text } from 'drizzle-orm/sqlite-core';
+
+export const users = sqliteTable('users_table', {
+  id: int().primaryKey({ autoIncrement: true }),
+  username: text().notNull(),
+  password: text().notNull(),
+  email: text(),
+  name: text().notNull(),
+  role: int().notNull(),
+  enabled: int().notNull().default(1),
+  createdAt: text('created_at')
+    .notNull()
+    .default(sql`(CURRENT_TIMESTAMP)`),
+  updatedAt: text('updated_at')
+    .notNull()
+    .default(sql`(CURRENT_TIMESTAMP)`)
+    .$onUpdate(() => sql`(CURRENT_TIMESTAMP)`),
+});
